chore(jans-cedarling): refactor move types to parse cedar-policy entities to file `entity_types.rs`
